+++ graphics/camera.h
@@ -7,13 +7,13 @@
 
 class camera {
 public:
-    camera(vec3 origin, vec3 forward, vec3 up, double fov, int horizontal_resolution, int vertical_resolution);
+    camera(vec3 origin, vec3 forward, vec3 up, float fov, int horizontal_resolution, int vertical_resolution);
 
     output_image capture_image(const hittable_list &scene) const;
 
 private:
     vec3 origin, forward, up, right;
-    double viewport_distance;
+    float viewport_distance;
     int horizontal_resolution, vertical_resolution;
 };
