+++ objects/sphere.h
@@ -6,14 +6,14 @@
 
 class sphere : public hittable {
 public:
-    sphere(double radius, const vec3 &center, const std::shared_ptr<material> &t);
+    sphere(float radius, const vec3 &center, const std::shared_ptr<material> &t);
     bool hit(ray& r, collision_info& hit_info) const override;
 
     aabb get_bounding_box() const override;
     vec3 centroid() const override;
 
 private:
-    double radius;
+    float radius;
     vec3 center;
 };
